Combine examples, formatting changes
